Refactor call duration and enhance group update logic
Updated call duration and modified Pygame display settings. Added functionality to call an external endpoint when updating the group number.

This had all been done in 2024 but hadn't been updated on GitHub.

New change in 2025, when running the program, it should automatically start on the projector display rather than the main display.